Restrict user searches to students (#55)
* Update OAuth Provider Class Location

A recent update to django-oauth-toolkit amongst other things moved the
Django OAuth Provider Classes from ext.rest_framework to
contrib.rest_framework. This caused an exception when running dreamjub
with the newest version of oauth-toolkit.

This commit fixes the problem by updating settings.py to use the
provider from the new source instead.

* Restrict user searches to students

After a request from IRC IT, we need to restrict users to students. This
commit updates the LDAP Query appropriatly.