Make 'contains' operator case-insensitive
Previously, the contains operator for the 'tl' users end point was case
sensitive, making it difficult to search for names without paying
specific attention to the case.

This commit updates the operator case insensitive, to enabling easier
searching for strings within names, both at the beginning and in the
middle of the name.