Update django_filters code
An update to django_filters caused the DjangoFilterBackend class to
switch namespace. This resulted in an error when attempting to start
dreamjub with newer versions of the package.

This commit fixes the issue by making an appropriate new import.
Furthermore, it adds the django_filters app to INSTALLED_APPS, ensuring
that it is loaded.