Switch to magic login links via email
After a request from IRC IT, it was neccessary to disable password-based
LDAP login. This commit updates the login form to instead allow login
via having links emailed.

This commit updates the login form, adding the new method as a default.
Furthermore, it adds a button to enable logging in via username and
password (through the standard backends) as well.