Fall 17 Fixup (#59)
* Switch to magic login links via email

After a request from IRC IT, it was neccessary to disable password-based
LDAP login. This commit updates the login form to instead allow login
via having links emailed.

This commit updates the login form, adding the new method as a default.
Furthermore, it adds a button to enable logging in via username and
password (through the standard backends) as well.

* Update django_filters code

An update to django_filters caused the DjangoFilterBackend class to
switch namespace. This resulted in an error when attempting to start
dreamjub with newer versions of the package.

This commit fixes the issue by making an appropriate new import.
Furthermore, it adds the django_filters app to INSTALLED_APPS, ensuring
that it is loaded.

* Update home page

This commit updates the home page to show more useful information about
the current user.

* Bugfix when local student exists

This commit fixes a bug that occured when a LocalStudent instance
existed.

* Bugfix for login urls

When logging in using certain urls the sesame token would cause a wrong
csrf token to be set, causing 403 errors on login.

This commit fixes the problem by introducing a /login/next page that the
user can be sent to.